requirements.txt
annotated-types==0.7.0
anyio==4.5.2
bcrypt==4.3.0
bson==0.5.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
dnspython==2.6.1
email-validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
fastapi-cli==0.0.7
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
passlib==1.7.4
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.44.0
typer==0.16.0
typing-extensions==4.13.2
urllib3==2.2.3
uvicorn==0.33.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
annotated-types==0.7.0
anyio==4.5.2
bcrypt==4.3.0
bson==0.5.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
dnspython==2.6.1
email-validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
fastapi-cli==0.0.7
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
passlib==1.7.4
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.44.0
typer==0.16.0
typing-extensions==4.13.2
urllib3==2.2.3
uvicorn==0.33.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
annotated-types==0.7.0
anyio==4.5.2
bcrypt==4.3.0
bson==0.5.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
dnspython==2.6.1
email-validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
fastapi-cli==0.0.7
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
passlib==1.7.4
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
PyJWT==2.9.0
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.44.0
typer==0.16.0
typing-extensions==4.13.2
urllib3==2.2.3
uvicorn==0.33.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
annotated-types==0.7.0
anyio==4.5.2
bcrypt==4.3.0
bson==0.5.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
dnspython==2.6.1
ecdsa==0.19.1
email-validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
fastapi-cli==0.0.7
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jinja2==3.1.6
jose==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
passlib==1.7.4
pyasn1==0.4.8
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
PyJWT==2.9.0
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.7
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.44.0
typer==0.16.0
typing-extensions==4.13.2
urllib3==2.2.3
uvicorn==0.33.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
